Move AS toHexString back to test
